Fixed QM scripts for Abyssea-Attohwa
