+++ scripts/zones/Abyssea-Attohwa/npcs/qm13.lua
@@ -1,26 +1,32 @@
 -----------------------------------
 -- Zone: Abyssea-Attohwa
 --  NPC: qm13 (???)
--- Spawns Smok
+-- Spawns Mielikki
 -- @pos ? ? ? 215
 -----------------------------------
-require("scripts/globals/keyitems");
 require("scripts/globals/status");
+
+-----------------------------------
+-- onTrade Action
+-----------------------------------
+
+function onTrade(player,npc,trade)
+--[[
+    if (trade:hasItemQty(3084,1) and trade:getItemCount() == 1) then -- Player has all the required items.
+        if (GetMobAction(17658273) == ACTION_NONE) then -- Mob not already spawned from this
+            SpawnMob(17658273):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
+            player:tradeComplete();
+        end
+    end
+]]
+end;
 
 -----------------------------------
 -- onTrigger Action
 -----------------------------------
 
 function onTrigger(player,npc)
---[[
-    if (GetMobAction(17658274) == ACTION_NONE) then -- NM not already spawned from this
-        if (player:hasKeyItem(HOLLOW_DRAGON_EYE)) then
-            player:startEvent(1022, HOLLOW_DRAGON_EYE); -- Ask if player wants to use KIs
-        else
-            player:startEvent(1023, HOLLOW_DRAGON_EYE); -- Do not ask, because player is missing at least 1.
-        end
-    end
-]]
+    player:startEvent(1010, 3084); -- Inform player what items they need.
 end;
 
 -----------------------------------
@@ -28,8 +34,8 @@ end;
 -----------------------------------
 
 function onEventUpdate(player,csid,option)
-    -- printf("CSID2: %u",csid);
-    -- printf("RESULT2: %u",option);
+    -- printf("CSID: %u",csid);
+    -- printf("RESULT: %u",option);
 end;
 
 -----------------------------------
@@ -39,8 +45,4 @@ end;
 function onEventFinish(player,csid,option)
     -- printf("CSID: %u",csid);
     -- printf("RESULT: %u",option);
-    if (csid == 1022 and option == 1) then
-        SpawnMob(17658274):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
-        player:delKeyItem(HOLLOW_DRAGON_EYE);
-    end
 end;
