+++ scripts/zones/Abyssea-Attohwa/npcs/qm17.lua
@@ -1,6 +1,6 @@
 -----------------------------------
 -- Zone: Abyssea-Attohwa
---  NPC: qm16 (???)
+--  NPC: qm17 (???)
 -- Spawns Itzpapalotl
 -- @pos ? ? ? 215
 -----------------------------------
