+++ scripts/zones/Abyssea-Attohwa/npcs/qm20.lua
@@ -1,6 +1,6 @@
 -----------------------------------
 -- Zone: Abyssea-Attohwa
---  NPC: qm19 (???)
+--  NPC: qm20 (???)
 -- Spawns Ulhuadshi
 -- @pos ? ? ? 215
 -----------------------------------
