+++ scripts/zones/Abyssea-Attohwa/npcs/qm23.lua
@@ -1,6 +1,6 @@
 -----------------------------------
 -- Zone: Abyssea-Attohwa
---  NPC: qm22 (???)
+--  NPC: qm23 (???)
 -- Spawns Titlacauan
 -- @pos ? ? ? 215
 -----------------------------------
