+++ scripts/zones/Abyssea-Attohwa/npcs/qm7.lua
@@ -1,7 +1,7 @@
 -----------------------------------
 -- Zone: Abyssea-Attohwa
 --  NPC: qm7 (???)
--- Spawns Kampe
+-- Spawns Svarbhanu
 -- @pos ? ? ? 215
 -----------------------------------
 require("scripts/globals/status");
@@ -12,9 +12,9 @@ require("scripts/globals/status");
 
 function onTrade(player,npc,trade)
 --[[
-    if (trade:hasItemQty(3079,1) and trade:getItemCount() == 1) then -- Player has all the required items.
-        if (GetMobAction(17658268) == ACTION_NONE) then -- Mob not already spawned from this
-            SpawnMob(17658268):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
+    if (trade:hasItemQty(3078,1) and trade:getItemCount() == 1) then -- Player has all the required items.
+        if (GetMobAction(17658267) == ACTION_NONE) then -- Mob not already spawned from this
+            SpawnMob(17658267):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
             player:tradeComplete();
         end
     end
@@ -26,7 +26,7 @@ end;
 -----------------------------------
 
 function onTrigger(player,npc)
-    player:startEvent(1010, 3079); -- Inform player what items they need.
+    player:startEvent(1010, 3078); -- Inform player what items they need.
 end;
 
 -----------------------------------
